Change long to int

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/controller/dto/recall/PrisonerRecallsResponse.kt
@@ -5,5 +5,5 @@ import io.swagger.v3.oas.annotations.media.Schema
 data class PrisonerRecallsResponse(
   val recalls: List<Recall>,
   @get:Schema(description = "Total number of recalls for the prisoner after NOMIS grouping")
-  val prisonerRecallTotal: Long,
+  val prisonerRecallTotal: Int,
 )
